+++ src/main/java/de/samply/notification/smtp/EmailService.java
@@ -13,6 +13,7 @@ import org.thymeleaf.context.Context;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 
 @Service
 @Slf4j
@@ -22,6 +23,7 @@ public class EmailService {
     private final JavaMailSender mailSender;
     private final TemplateEngine templateEngine;
     private final EmailTemplates emailTemplates;
+
 
     public EmailService(
             @Value(ProjectManagerConst.PROJECT_MANAGER_EMAIL_FROM_SV) String emailFrom,
@@ -33,24 +35,36 @@ public class EmailService {
         this.templateEngine = templateEngine;
         this.emailTemplates = emailTemplates;
     }
-    public void sendEmail(@NotNull String email, @NotNull ProjectRole role, @NotNull EmailTemplateType type){
+
+    public void sendEmail(@NotNull String email, @NotNull ProjectRole role, @NotNull EmailTemplateType type) throws EmailServiceException {
         sendEmail(email, role, type, new HashMap<>());
     }
 
-    public void sendEmail(@NotNull String email, @NotNull ProjectRole role, @NotNull EmailTemplateType type, Map<String, String> keyValues){
+    public void sendEmail(@NotNull String email, @NotNull ProjectRole role, @NotNull EmailTemplateType type, Map<String, String> keyValues) throws EmailServiceException {
         SimpleMailMessage message = new SimpleMailMessage();
-        message.setTo(email);
-        message.setFrom(emailFrom);
-        message.setSubject("TODO: REPLACE ME"); //TODO
-        message.setText(createEmailBody(role, type, keyValues));
-        mailSender.send(message);
+        Optional<MessageSubject> messageSubject = createEmailMessageAndSubject(role, type, keyValues);
+        if (messageSubject.isPresent()) {
+            message.setTo(email);
+            message.setFrom(emailFrom);
+            message.setSubject(messageSubject.get().subject());
+            message.setText(messageSubject.get().message());
+            mailSender.send(message);
+        } else {
+            throw new EmailServiceException("Template not found for " + type.name() + " of role " + role.name());
+        }
+
     }
 
-    private String createEmailBody(ProjectRole role, EmailTemplateType type, Map<String, String> keyValues) {
-        return templateEngine.process(emailTemplates.getTemplate(type, role).get(), createContext(keyValues));
+    private Optional<MessageSubject> createEmailMessageAndSubject(ProjectRole role, EmailTemplateType type, Map<String, String> keyValues) {
+        Optional<TemplateSubject> template = emailTemplates.getTemplateAndSubject(type, role);
+        if (template.isPresent()) {
+            String message = templateEngine.process(template.get().template(), createContext(keyValues));
+            return Optional.of(new MessageSubject(message, template.get().subject()));
+        }
+        return Optional.empty();
     }
 
-    private String createEmailBody(String template, Map<String, String> keyValues) {
+    private String createEmailMessageAndSubject(String template, Map<String, String> keyValues) {
         return templateEngine.process(template, createContext(keyValues));
     }